Create Construct BST from inorder and preorder
As discussed above the preorder traversal follows Root -> Left -> Right order, that makes it very convenient to construct the tree from its root.

Let's do it. The first element in the preorder list is a root. This root splits inorder list into left and right subtrees. Now one have to pop up the root from preorder list since it's already used as a tree node and then repeat the step above for the left and right subtrees.

https://leetcode.com/problems/construct-binary-tree-from-preorder-and-inorder-traversal/solution/